Fixed bug with detection time. support for exposure time, gain and cpus being returned as part of preview

+++ Website/Models/DetectionLoader.cs
@@ -84,7 +84,7 @@ public class DetectionResult {
     public DetectionResult() {
         tracking_data = new List<RawTrackingData>();
     }
-    public float posix_time {get; set;}
+    public double posix_time {get; set;}
     public float mean_speed {get; set;}
     public DetectionDirection direction {get; set;}
     public float sd {get; set;}
